fix upadate profile data in userinfo component, delete userslice
