+++ src/redux/store.js
@@ -11,7 +11,6 @@ import {
   REGISTER,
 } from 'redux-persist';
 import { authReducer } from './auth/slice';
-import { userReducer } from './user/userSlice';
 import { projectReducer } from './project/projectSlice';
 
 const persistConfig = {
@@ -25,7 +24,6 @@ const persistedAuthReducer = persistReducer(persistConfig, authReducer);
 export const store = configureStore({
   reducer: {
     auth: persistedAuthReducer,
-    user: userReducer,
     project: projectReducer,
   },
   middleware: getDefaultMiddleware =>
